Forgot replacing an AddEdge() with the _addEdge() (without validation) in DichotomizeLayer in Network processor Added TryGetValue() in RandomWalkNodeAttrFPT to avoid a possible exception being thrown (for missing node attr values) Different kind of loop in ShortestPathNodeAttributeDistances(): faster

+++ Threadle.Core/Analysis/Functions.cs
@@ -487,7 +487,7 @@ namespace Threadle.Core.Analysis
                 int k = layerTwoMode.GetHyperedgeNodeIds(name).Length;
                 if (k >= 2)
                 {
-                    long weight = k * (k - 1) / 2;
+                    long weight = (long)k * (k - 1) / 2;
                     validNames.Add(name);
                     weights.Add(weight);
                     totalWeight += weight;
